backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
tensorflow-cpu==2.15.0
python-dotenv==1.0.1
requests==2.31.0